[Clean] Change directory structures of stories
