+++ stories/0-vertical-simple/shared/common.css
