+++ stories/0-vertical-simple/shared/index.ts
@@ -1,0 +1,1 @@
+export { default as styles } from "./common.css";
