+++ stories/1-vertical-nested/index.ts
@@ -1,0 +1,2 @@
+export { DynamicComponent } from "./dynamic.component";
+export { StaticComponent } from "./static.component";
