+++ stories/1-vertical-nested/shared/common.css
