+++ stories/1-vertical-nested/shared/index.ts
@@ -1,0 +1,1 @@
+export { default as styles } from "./common.css";
