+++ stories/2-vertical-tree/index.ts
@@ -1,0 +1,2 @@
+export { DynamicComponent } from "./dynamic.component";
+export { StaticComponent } from "./static.component";
