+++ stories/2-vertical-tree/shared/index.ts
@@ -1,0 +1,1 @@
+export { default as styles } from "./common.css";
